bookstore/requirements.txt
alipay==0.7.4
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crcmod==1.7
cryptography==41.0.3
Django==4.2.4
django-aliyun-oss2-storage==0.1.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-crontab==0.7.1
django-filter==23.2
django-js-asset==2.1.0
django-redis==5.3.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
idna==3.4
jmespath==0.10.0
kombu==5.3.1
oss2==2.18.1
Pillow==10.0.0
prompt-toolkit==3.0.39
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
PyJWT==1.7.1
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-alipay-sdk==3.3.0
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.0
requests==2.31.0
simpleui==4.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
